+++ examples/chart_new.rs
@@ -3,7 +3,6 @@ use rs_backtester::backtester_new::Backtest_arc;
 use rs_backtester::charts::{plot_arc, PlotConfig};
 use rs_backtester::datas::Data;
 use rs_backtester::strategies::sma_cross_arc;
-use rs_backtester::chart_test::show_equity_curve;
 
 pub fn main() -> Result<(),Box<dyn Error>> {
     //example to plot backtesting results and how to change chart configuration
@@ -17,6 +16,5 @@ pub fn main() -> Result<(),Box<dyn Error>> {
     plot_config.display_networth = true;
 
     plot_arc(&backtest,plot_config,"plot.png")?;
-    show_equity_curve(data.datetime.clone().into_iter().map(|dt|dt.date_naive()).collect(),data.close.clone());
     Ok(())
 }