+++ examples/strategy_new.rs
@@ -1,15 +1,13 @@
 use std::error::Error;
 use std::sync::Arc;
 use rs_backtester::datas::Data;
-use rs_backtester::strategies::{Strategy, Strategy_arc};
+use rs_backtester::strategies::{Strategy_arc};
 use rs_backtester::orders::Order::{BUY,SHORTSELL,NULL};
 extern crate rand;
 use rand::thread_rng;
 use rand::seq::SliceRandom;
-use rs_backtester::backtester::{Backtest, Commission};
 use rs_backtester::backtester_new::Backtest_arc;
 use rs_backtester::metrics::report_horizontal_arc;
-use rs_backtester::report::report;
 
 
 pub fn main() -> Result<(),Box<dyn Error>>{
@@ -17,7 +15,7 @@ pub fn main() -> Result<(),Box<dyn Error>>{
     let quotes = Data::new_from_yahoo_arc("PLTR","1d","6mo")?;
 
     pub fn random_strategy(quotes:Arc<Data>)->Strategy_arc{
-        let length = quotes.timestamps().len();
+        let length = quotes.datetime.len();
         let mut choices = vec![NULL;length];
         let name = "random strategy".to_string();
         let indicator = Some(vec![vec![-1.;length]]);
@@ -27,8 +25,8 @@ pub fn main() -> Result<(),Box<dyn Error>>{
             choices[i] = *rnd_orders.choose(&mut rng).unwrap();
         }
         Strategy_arc{
-            name:name,
-            choices:choices,
+            name,
+            choices,
             indicator,
             data:quotes.clone(),
         }