+++ src/backtester_new.rs
@@ -4,11 +4,10 @@ use crate::broker;
 use crate::broker::Broker;
 use crate::config::{Config, CONFIG};
 use crate::metrics::Metrics;
-use crate::strategies::{Strategy,Strategy_arc};
-use crate::datas::Data;
+use crate::strategies::{Strategy_arc};
 use crate::trades::trade_indices_from_broker;
 use crate::utilities::{SerializeAsCsv, write_combined_csv};
-
+/*
 pub struct Backtest{
     pub(super) quotes: Data,
     strategy: Strategy,
@@ -57,7 +56,7 @@ impl Backtest{
         println!(" execution time = {:?}\x1b[0m",self.local_config.clone().unwrap().execution_time);
     }
 }
-
+*/
 pub struct Backtest_arc{
     pub strategy: Strategy_arc,
     pub(crate) broker: Broker,
@@ -68,13 +67,14 @@ pub struct Backtest_arc{
 impl crate::backtester_new::Backtest_arc {
     pub fn new(strategy: Strategy_arc, initial_account: f64) -> Self {
         let local_config = {
-            let guard = CONFIG.read().unwrap().clone().unwrap_or_else(Config::default);
+            //let guard = CONFIG.read().unwrap().clone().unwrap_or_else(Config::default);
+            let guard = CONFIG.read().unwrap().clone().unwrap_or_default();
             guard.clone()
         };
         let start = Instant::now();
         let broker = broker::calculate_arc(&strategy, initial_account);
         let duration = start.elapsed();
-        println!("\x1b[34mBacktesting (ARC) for {} calculated in {:?}\x1b[0m", strategy.name, duration);
+        println!("\x1b[34mBacktesting for {} calculated in {:?}\x1b[0m", strategy.name, duration);
         let mut metrics = Metrics::default();
         broker.calculate_metrics(&mut metrics);
         let trades = trade_indices_from_broker(&broker);
@@ -82,11 +82,11 @@ impl crate::backtester_new::Backtest_arc {
         //adds TradesIndices to Metrics
         broker.trade_indices(&mut metrics);
         let duration = start.elapsed();
-        println!("\x1b[34mBacktesting and metrics (ARC) for { } calculated in {:?}\x1b[0m", strategy.name, duration);
+        println!("\x1b[34mBacktesting and metrics for { } calculated in {:?}\x1b[0m", strategy.name, duration);
         crate::backtester_new::Backtest_arc {
-            strategy: strategy,
-            broker: broker,
-            metrics: metrics,
+            strategy,
+            broker,
+            metrics,
             local_config: Some(local_config),
         }
     }