+++ src/lib.rs
@@ -3,7 +3,6 @@ pub mod errors;
 pub mod charts;
 pub mod orders;
 pub mod strategies;
-pub mod backtester;
 pub mod ta;
 pub mod report;
 pub mod trades;
@@ -12,6 +11,4 @@ pub mod broker;
 pub mod metrics;
 pub mod backtester_new;
 pub mod config;
-pub mod chart_test;
-
 pub use errors::{Result};