+++ src/strategies.rs
@@ -4,11 +4,11 @@ use crate::datas::Data;
 use crate::orders::Order;
 use crate::orders::Order::{BUY,SHORTSELL,NULL};
 //use std::error::Error;
-use crate::ta::{Indicator,sma,rsi, Indicator_arc};
+use crate::ta::{sma,rsi, Indicator_arc};
 use serde::{Serialize};
 //use serde::de::Unexpected::Str;
 use std::sync::Arc;
-
+/*
 /// Struct to hold vector of choices and indicators<BR>
 /// There is no specific constructor<BR>
 /// Need to be created via a user-defined function which return a Strategy
@@ -17,7 +17,10 @@ pub struct Strategy{
     pub name:String,
     pub choices:Vec<Order>,
     pub indicator:Option<Vec<Vec<f64>>>,
-}
+}*/
+/// Struct to hold vector of choices and indicators<BR>
+/// There is no specific constructor<BR>
+/// Need to be created via a user-defined function which return a Strategy
 #[derive(Clone, Serialize)]
 pub struct Strategy_arc{
     pub name:String,
@@ -26,7 +29,7 @@ pub struct Strategy_arc{
     pub data:Arc<Data>,
 }
 
-
+/*
 impl Strategy{
     pub fn choices(&self)->Vec<Order>{
         return self.choices.clone();
@@ -93,15 +96,17 @@ impl Strategy{
             indicator,
         }
     }
-}
+}*/
 
 impl Strategy_arc{
     ///invert strategy (LONG->SHORT and viceversa)
     pub fn invert(&self) ->Self{
-        let length = self.choices.len();
         let mut inv_choices = self.choices.clone();
-        for i in 0..length{
+        /*for i in 0..length{
             if self.choices[i]==BUY { inv_choices[i]=SHORTSELL} else if self.choices[i]==SHORTSELL { inv_choices[i]=BUY}
+        }*/
+        for (choices, inv_choices) in self.choices.iter().zip(inv_choices.iter_mut()){
+            if *choices==BUY {*inv_choices=SHORTSELL} else if *choices==SHORTSELL{ *inv_choices=BUY };
         }
         let indicator = self.indicator.clone();
         Strategy_arc{
@@ -111,11 +116,15 @@ impl Strategy_arc{
             data:self.data.clone(),
         }
     }
+    ///transform current strategy in long only
     pub fn long_only(&self) ->Self{
-        let length = self.choices.len();
         let mut long_choices = self.choices.clone();
+        /*
         for i in 0..length{
             if self.choices[i]==SHORTSELL { long_choices[i]=NULL}
+        }*/
+        for (choices,long_choices) in self.choices.iter().zip(long_choices.iter_mut()){
+            if *choices==SHORTSELL{*long_choices=NULL}
         }
         let indicator = self.indicator.clone();
         Strategy_arc{
@@ -124,12 +133,17 @@ impl Strategy_arc{
             indicator,
             data:self.data.clone(),
         }
-    }
+    }///
+    ///transform current strategy in short only
     pub fn short_only(&self) ->Self{
-        let length = self.choices.len();
+        //let length = self.choices.len();
         let mut short_choices = self.choices.clone();
+        /*
         for i in 0..length{
             if self.choices[i]==BUY { short_choices[i]=NULL}
+        }*/
+        for (choice,short_choices) in self.choices.iter().zip(short_choices.iter_mut()){
+            if *choice==BUY{*short_choices=NULL}
         }
         let indicator = self.indicator.clone();
         Strategy_arc{
@@ -144,6 +158,7 @@ impl Strategy_arc{
         let mut change_count = 0;
         let mut new_choices = self.choices.clone();
         let indicator=self.indicator.clone();
+        /*
         for i in 1..self.choices.len(){
             if self.choices[i]!=self.choices[i-1]{
                 change_count += 1;
@@ -152,7 +167,19 @@ impl Strategy_arc{
                 new_choices[i]=Order::NULL;
             }
             else { new_choices[i]=self.choices[i] }
+        }*/
+        for (i, window) in self.choices.windows(2).enumerate() {
+            if window[0] != window[1] {
+                change_count += 1;
+            }
+
+            new_choices[i + 1] = if change_count < 2 {
+                Order::NULL
+            } else {
+                window[1]
+            };
         }
+
         Strategy_arc{
             name:self.name.clone()+"_skip",
             choices:new_choices,
@@ -161,7 +188,7 @@ impl Strategy_arc{
         }
     }
 }
-
+/*
 ///Returns Buy and Hold Strategy
 pub fn buy_n_hold(quotes:Data)->Strategy{
     let length = quotes.timestamps().len();
@@ -173,18 +200,18 @@ pub fn buy_n_hold(quotes:Data)->Strategy{
         choices:choices,
         indicator,
     }
-}
+}*/
 ///Returns Buy and Hold Strategy
-/// todo! buy and hold start from 3rd+ period (1 to see the data, 1 to send order, 1+ to execute)
+/// todo! buy and hold start from 3rd+ period (1 to see the data (order is on change of stance), 1 to send order, 1+ to execute)
 pub fn buy_n_hold_arc(quotes:Arc<Data>)->Strategy_arc{
-    let length = quotes.timestamps().len();
+    let length = quotes.datetime.len();
     let mut choices = vec![BUY;length];
     let name = "buy_and_hold".to_string();
     let indicator = Some(vec![vec![-1.;length]]);
     choices[0]=NULL;
     Strategy_arc{
-        name:name,
-        choices:choices,
+        name,
+        choices,
         indicator,
         data:quotes.clone(),
     }
@@ -192,29 +219,32 @@ pub fn buy_n_hold_arc(quotes:Arc<Data>)->Strategy_arc{
 
 ///Returns the opposite of a Buy and Hold Strategy:
 /// start by shortselling and keep the short position open to the end
-pub fn short_n_hold(quotes:Data)->Strategy{
-    let length = quotes.timestamps().len();
+pub fn short_n_hold_arc(quotes:Arc<Data>)->Strategy_arc{
+    let length = quotes.datetime.len();
     let choices = vec![SHORTSELL;length];
     let name = "short and hold".to_string();
     let indicator = Some(vec![vec![-1.;length]]);
-    Strategy{
-        name:name,
-        choices:choices,
+    Strategy_arc{
+        name,
+        choices,
         indicator,
+        data:quotes.clone(),
     }
 }
 ///Returns a Strategy which does exactly nothing (i.e. always stays out of the market)
-pub fn do_nothing(quotes:Data)->Strategy{
-    let length = quotes.timestamps().len();
+pub fn do_nothing_arc(quotes:Arc<Data>)->Strategy_arc{
+    let length = quotes.datetime.len();
     let choices = vec![NULL;length];
     let name = "do nothing".to_string();
     let indicator = Some(vec![vec![-1.;length]]);
-    Strategy{
-        name:name,
-        choices:choices,
+    Strategy_arc{
+        name,
+        choices,
         indicator,
+        data:quotes.clone(),
     }
 }
+/*
 ///Returns a Simple Moving Average Strategy with a user specified time-period
 pub fn simple_sma(quotes:Data, period:usize) ->Strategy{
     let sma = sma(&quotes,period);
@@ -236,31 +266,37 @@ pub fn simple_sma(quotes:Data, period:usize) ->Strategy{
         choices:choices,
         indicator,
     }
-}
+}*/
 ///Returns a Simple Moving Average Strategy with a user specified time-period
-pub fn simple_sma_arc(quotes:Arc<Data>, period:usize) ->Strategy_arc{
+pub fn sma_arc(quotes:Arc<Data>, period:usize) ->Strategy_arc{
     let sma = sma(&quotes,period);
     let indicator = Indicator_arc{indicator:sma,quotes:quotes.clone()};
-    let length = indicator.quotes.timestamps().len();
+    let length = indicator.quotes.datetime.len();
     let mut choices = vec![NULL;length];
-    for i in 0..length{
+    /*for i in 0..length{
         if indicator.indicator[i]!=-1.{
-            if indicator.indicator[i]<=indicator.quotes.close()[i]{
+            if indicator.indicator[i]<=indicator.quotes.close[i]{
                 choices[i] = BUY;
-            }else if indicator.indicator[i]>indicator.quotes.close()[i]{
+            }else if indicator.indicator[i]>indicator.quotes.close[i]{
                 choices[i] = SHORTSELL}
         }
+    }*/
+    for ((ind, close), choice) in indicator.indicator.iter().zip(&indicator.quotes.close).zip(choices.iter_mut())
+    {
+        if *ind != -1. {
+            *choice = if *ind <= *close {BUY} else {SHORTSELL};
+        }
     }
-    let name = format!("simple_sma_{}",period);
+    let name = format!("sma_{}",period);
     let indicator = Some(vec![indicator.indicator]);
     Strategy_arc{
-        name:name,
-        choices:choices,
+        name,
+        choices,
         indicator,
         data:quotes.clone(),
     }
 }
-
+/*
 ///Returns a Simple Moving Average Crossing Strategy (i.e. goes long when SMA short crosses SMA long and shortsells otherwise)<BR>
 ///User can specify both time-periods (short and long, with short first)
 pub fn sma_cross(quotes:Data, short_period:usize, long_period:usize)->Strategy{
@@ -284,32 +320,37 @@ pub fn sma_cross(quotes:Data, short_period:usize, long_period:usize)->Strategy{
         choices:choices,
         indicator:indicator,
     }
-}
-
+}*/
+///Returns a Simple Moving Average Crossing Strategy (i.e. goes long when SMA short crosses SMA long and shortsells otherwise)<BR>
+///User can specify both time-periods (short and long, with short first)
 pub fn sma_cross_arc(quotes:Arc<Data>, short_period:usize, long_period:usize)->Strategy_arc{
     if short_period >= long_period {panic!("Error: short SMA parameter should be shorter than long SMA parameter");}
     let sma_short = sma(&quotes, short_period);
     let sma_long = sma(&quotes, long_period);
     let ind_short = Indicator_arc{indicator:sma_short,quotes:quotes.clone()};
     let ind_long = Indicator_arc{indicator:sma_long, quotes:quotes.clone()};
-    let length = ind_short.quotes.timestamps().len();
+    let length = ind_short.quotes.datetime.len();
     let mut choices = vec![NULL;length];
+    for ((ind_long,ind_short),choices) in ind_long.indicator.iter().zip(ind_short.indicator.iter()).zip(choices.iter_mut()){
+        if *ind_long!=-1.{ *choices = if *ind_short>*ind_long {BUY} else { SHORTSELL } }
+    }
+    /*
     for i in 0..length{
         if ind_long.indicator[i]!=-1.{
             if ind_short.indicator[i]>ind_long.indicator[i]{choices[i]=BUY}
             else {choices[i]=SHORTSELL};
         }
-    }
+    }*/
     let name=format!("sma_cross_{}_{}",short_period,long_period);
     let indicator = Some(vec![ind_short.indicator,ind_long.indicator]);
     Strategy_arc{
-        name:name,
-        choices:choices,
-        indicator:indicator,
+        name,
+        choices,
+        indicator,
         data:quotes,
     }
 }
-
+/*
 ///Returns a Relative Strength Index Strategy (i.e. goes short if RSI > 70, long when RSI < 30, and stay out of market elsewhere)
 pub fn rsi_strategy(quotes:Data, period:usize)->Strategy{
     let rsi = rsi(&quotes,period);
@@ -329,18 +370,21 @@ pub fn rsi_strategy(quotes:Data, period:usize)->Strategy{
         choices,
         indicator,
     }
-}
+}*/
 ///Returns a Relative Strength Index Strategy (i.e. goes short if RSI > 70, long when RSI < 30, and stay out of market elsewhere)
 pub fn rsi_strategy_arc(quotes:Arc<Data>, period:usize)->Strategy_arc{
     let rsi = rsi(&quotes,period);
     let indicator = Indicator_arc{indicator:rsi,quotes:quotes.clone()};
-    let length = indicator.quotes.timestamps().len();
+    let length = indicator.quotes.datetime.len();
     let mut choices = vec![NULL;length];
-    for i in 0..length{
+    /*for i in 0..length{
         if indicator.indicator[i]!=-1.{
             if indicator.indicator[i]>70.{choices[i]=SHORTSELL}
             else if indicator.indicator[i]<30. {choices[i]=BUY}
         }
+    }*/
+    for (ind,choice) in indicator.indicator.iter().zip(choices.iter_mut()){
+        if *ind !=-1.{if *ind > 70.{*choice = SHORTSELL} else if *ind < 30. {*choice = BUY}}
     }
     let name = format!("rsi_{}",period);
     let indicator=Some(vec![indicator.indicator]);