+++ CHANGELOG.md
@@ -6,8 +6,11 @@
 - simplification of csv generation (with the exception of backtesting under rework) with possibility to combine in a single csv more granular csv output
 - rework of engines (now calculations are made at backtest initialization and stored in Metrics and not in multiple places)
 - added computation timing printout
-- improved performance
-- reworks of backtesting and strategies
+- improved performance (more than 4x on backtesting)
+- rework of backtesting and strategies
+- rework of reports
+- rework of examples
+- rework of tests
 
 ## [0.1.3]
 - Possibility to display the list of trades produced by a strategy